refactor: rename response-types.ts to data-types.ts to reflect broader usage
