--- src/lib/response-types.ts
@@ -1,8 +1,0 @@
-export type PlugConfiguration = {
-    plug_url:string,
-    natural_factor: number,
-    local_latitude: number,
-    local_longitude: number,
-    natural_latitude: number,
-    natural_longitude: number
-};
